Convert Media Management settings to TypeScript (Sonarr/Sonarr@27f81117e)
