--- frontend/src/Helpers/getDisplayName.js
@@ -1,3 +1,0 @@
-export default function getDisplayName(Component) {
-  return Component.displayName || Component.name || 'Component';
-}
