--- frontend/src/Settings/MediaManagement/MediaManagement.js
@@ -1,536 +1,0 @@
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import Alert from 'Components/Alert';
-import FieldSet from 'Components/FieldSet';
-import Form from 'Components/Form/Form';
-import FormGroup from 'Components/Form/FormGroup';
-import FormInputGroup from 'Components/Form/FormInputGroup';
-import FormLabel from 'Components/Form/FormLabel';
-import LoadingIndicator from 'Components/Loading/LoadingIndicator';
-import PageContent from 'Components/Page/PageContent';
-import PageContentBody from 'Components/Page/PageContentBody';
-import { inputTypes, kinds, sizes } from 'Helpers/Props';
-import RootFolders from 'RootFolder/RootFolders';
-import SettingsToolbar from 'Settings/SettingsToolbar';
-import translate from 'Utilities/String/translate';
-import NamingConnector from './Naming/NamingConnector';
-import AddRootFolder from './RootFolder/AddRootFolder';
-
-const episodeTitleRequiredOptions = [
-  {
-    key: 'always',
-    get value() {
-      return translate('Always');
-    }
-  },
-  {
-    key: 'bulkSeasonReleases',
-    get value() {
-      return translate('OnlyForBulkSeasonReleases');
-    }
-  },
-  {
-    key: 'never',
-    get value() {
-      return translate('Never');
-    }
-  }
-];
-
-const rescanAfterRefreshOptions = [
-  {
-    key: 'always',
-    get value() {
-      return translate('Always');
-    }
-  },
-  {
-    key: 'afterManual',
-    get value() {
-      return translate('AfterManualRefresh');
-    }
-  },
-  {
-    key: 'never',
-    get value() {
-      return translate('Never');
-    }
-  }
-];
-
-const downloadPropersAndRepacksOptions = [
-  {
-    key: 'preferAndUpgrade',
-    get value() {
-      return translate('PreferAndUpgrade');
-    }
-  },
-  {
-    key: 'doNotUpgrade',
-    get value() {
-      return translate('DoNotUpgradeAutomatically');
-    }
-  },
-  {
-    key: 'doNotPrefer',
-    get value() {
-      return translate('DoNotPrefer');
-    }
-  }
-];
-
-const fileDateOptions = [
-  {
-    key: 'none',
-    get value() {
-      return translate('None');
-    }
-  },
-  {
-    key: 'localAirDate',
-    get value() {
-      return translate('LocalAirDate');
-    }
-  },
-  {
-    key: 'utcAirDate',
-    get value() {
-      return translate('UtcAirDate');
-    }
-  }
-];
-
-class MediaManagement extends Component {
-
-  //
-  // Render
-
-  render() {
-    const {
-      advancedSettings,
-      isFetching,
-      error,
-      settings,
-      hasSettings,
-      isWindows,
-      onInputChange,
-      onSavePress,
-      ...otherProps
-    } = this.props;
-
-    return (
-      <PageContent title={translate('MediaManagementSettings')}>
-        <SettingsToolbar
-          advancedSettings={advancedSettings}
-          {...otherProps}
-          onSavePress={onSavePress}
-        />
-
-        <PageContentBody>
-          <NamingConnector />
-
-          {
-            isFetching ?
-              <FieldSet legend={translate('NamingSettings')}>
-                <LoadingIndicator />
-              </FieldSet> : null
-          }
-
-          {
-            !isFetching && error ?
-              <FieldSet legend={translate('NamingSettings')}>
-                <Alert kind={kinds.DANGER}>
-                  {translate('MediaManagementSettingsLoadError')}
-                </Alert>
-              </FieldSet> : null
-          }
-
-          {
-            hasSettings && !isFetching && !error ?
-              <Form
-                id="mediaManagementSettings"
-                {...otherProps}
-              >
-                {
-                  advancedSettings ?
-                    <FieldSet legend={translate('Folders')}>
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('CreateEmptySiteFolders')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="createEmptySeriesFolders"
-                          helpText={translate('CreateEmptySiteFoldersHelpText')}
-                          onChange={onInputChange}
-                          {...settings.createEmptySeriesFolders}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('DeleteEmptyFolders')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="deleteEmptyFolders"
-                          helpText={translate('DeleteEmptyFoldersHelpText')}
-                          onChange={onInputChange}
-                          {...settings.deleteEmptyFolders}
-                        />
-                      </FormGroup>
-                    </FieldSet> : null
-                }
-
-                {
-                  advancedSettings ?
-                    <FieldSet
-                      legend={translate('Importing')}
-                    >
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.SMALL}
-                      >
-                        <FormLabel>{translate('EpisodeTitleRequired')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.SELECT}
-                          name="episodeTitleRequired"
-                          helpText={translate('EpisodeTitleRequiredHelpText')}
-                          values={episodeTitleRequiredOptions}
-                          onChange={onInputChange}
-                          {...settings.episodeTitleRequired}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('SkipFreeSpaceCheck')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="skipFreeSpaceCheckWhenImporting"
-                          helpText={translate('SkipFreeSpaceCheckHelpText')}
-                          onChange={onInputChange}
-                          {...settings.skipFreeSpaceCheckWhenImporting}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('MinimumFreeSpace')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.NUMBER}
-                          unit='MB'
-                          name="minimumFreeSpaceWhenImporting"
-                          helpText={translate('MinimumFreeSpaceHelpText')}
-                          onChange={onInputChange}
-                          {...settings.minimumFreeSpaceWhenImporting}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('UseHardlinksInsteadOfCopy')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="copyUsingHardlinks"
-                          helpText={translate('CopyUsingHardlinksHelpText')}
-                          helpTextWarning={translate('CopyUsingHardlinksHelpTextWarning')}
-                          onChange={onInputChange}
-                          {...settings.copyUsingHardlinks}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('ImportUsingScript')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="useScriptImport"
-                          helpText={translate('ImportUsingScriptHelpText')}
-                          onChange={onInputChange}
-                          {...settings.useScriptImport}
-                        />
-                      </FormGroup>
-
-                      {
-                        settings.useScriptImport.value ?
-                          <FormGroup
-                            advancedSettings={advancedSettings}
-                            isAdvanced={true}
-                          >
-                            <FormLabel>{translate('ImportScriptPath')}</FormLabel>
-
-                            <FormInputGroup
-                              type={inputTypes.PATH}
-                              includeFiles={true}
-                              name="scriptImportPath"
-                              helpText={translate('ImportScriptPathHelpText')}
-                              onChange={onInputChange}
-                              {...settings.scriptImportPath}
-                            />
-                          </FormGroup> : null
-                      }
-
-                      <FormGroup size={sizes.MEDIUM}>
-                        <FormLabel>{translate('ImportExtraFiles')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="importExtraFiles"
-                          helpText={translate('ImportExtraFilesHelpText')}
-                          onChange={onInputChange}
-                          {...settings.importExtraFiles}
-                        />
-                      </FormGroup>
-
-                      {
-                        settings.importExtraFiles.value ?
-                          <FormGroup
-                            advancedSettings={advancedSettings}
-                            isAdvanced={true}
-                          >
-                            <FormLabel>{translate('ImportExtraFiles')}</FormLabel>
-
-                            <FormInputGroup
-                              type={inputTypes.TEXT}
-                              name="extraFileExtensions"
-                              helpTexts={[
-                                translate('ExtraFileExtensionsHelpText'),
-                                translate('ExtraFileExtensionsHelpTextsExamples')
-                              ]}
-                              onChange={onInputChange}
-                              {...settings.extraFileExtensions}
-                            />
-                          </FormGroup> : null
-                      }
-                    </FieldSet> : null
-                }
-
-                <FieldSet
-                  legend={translate('FileManagement')}
-                >
-                  <FormGroup size={sizes.MEDIUM}>
-                    <FormLabel>{translate('UnmonitorDeletedEpisodes')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.CHECK}
-                      name="autoUnmonitorPreviouslyDownloadedEpisodes"
-                      helpText={translate('UnmonitorDeletedEpisodesHelpText')}
-                      onChange={onInputChange}
-                      {...settings.autoUnmonitorPreviouslyDownloadedEpisodes}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                    size={sizes.MEDIUM}
-                  >
-                    <FormLabel>{translate('DownloadPropersAndRepacks')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.SELECT}
-                      name="downloadPropersAndRepacks"
-                      helpTexts={[
-                        translate('DownloadPropersAndRepacksHelpText'),
-                        translate('DownloadPropersAndRepacksHelpTextCustomFormat')
-                      ]}
-                      helpTextWarning={
-                        settings.downloadPropersAndRepacks.value === 'doNotPrefer' ?
-                          translate('DownloadPropersAndRepacksHelpTextWarning') :
-                          undefined
-                      }
-                      values={downloadPropersAndRepacksOptions}
-                      onChange={onInputChange}
-                      {...settings.downloadPropersAndRepacks}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                    size={sizes.MEDIUM}
-                  >
-                    <FormLabel>{translate('AnalyseVideoFiles')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.CHECK}
-                      name="enableMediaInfo"
-                      helpText={translate('AnalyseVideoFilesHelpText')}
-                      onChange={onInputChange}
-                      {...settings.enableMediaInfo}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                  >
-                    <FormLabel>{translate('RescanSiteFolderAfterRefresh')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.SELECT}
-                      name="rescanAfterRefresh"
-                      helpText={translate('RescanAfterRefreshHelpText')}
-                      helpTextWarning={translate('RescanAfterRefreshHelpTextWarning')}
-                      values={rescanAfterRefreshOptions}
-                      onChange={onInputChange}
-                      {...settings.rescanAfterRefresh}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                  >
-                    <FormLabel>{translate('ChangeFileDate')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.SELECT}
-                      name="fileDate"
-                      helpText={translate('ChangeFileDateHelpText')}
-                      values={fileDateOptions}
-                      onChange={onInputChange}
-                      {...settings.fileDate}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                  >
-                    <FormLabel>{translate('RecyclingBin')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.PATH}
-                      name="recycleBin"
-                      helpText={translate('RecyclingBinHelpText')}
-                      onChange={onInputChange}
-                      {...settings.recycleBin}
-                    />
-                  </FormGroup>
-
-                  <FormGroup
-                    advancedSettings={advancedSettings}
-                    isAdvanced={true}
-                  >
-                    <FormLabel>{translate('RecyclingBinCleanup')}</FormLabel>
-
-                    <FormInputGroup
-                      type={inputTypes.NUMBER}
-                      name="recycleBinCleanupDays"
-                      helpText={translate('RecyclingBinCleanupHelpText')}
-                      helpTextWarning={translate('RecyclingBinCleanupHelpTextWarning')}
-                      min={0}
-                      onChange={onInputChange}
-                      {...settings.recycleBinCleanupDays}
-                    />
-                  </FormGroup>
-                </FieldSet>
-
-                {
-                  advancedSettings && !isWindows ?
-                    <FieldSet
-                      legend={translate('Permissions')}
-                    >
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                        size={sizes.MEDIUM}
-                      >
-                        <FormLabel>{translate('SetPermissions')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.CHECK}
-                          name="setPermissionsLinux"
-                          helpText={translate('SetPermissionsLinuxHelpText')}
-                          helpTextWarning={translate('SetPermissionsLinuxHelpTextWarning')}
-                          onChange={onInputChange}
-                          {...settings.setPermissionsLinux}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                      >
-                        <FormLabel>{translate('ChmodFolder')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.UMASK}
-                          name="chmodFolder"
-                          helpText={translate('ChmodFolderHelpText')}
-                          helpTextWarning={translate('ChmodFolderHelpTextWarning')}
-                          onChange={onInputChange}
-                          {...settings.chmodFolder}
-                        />
-                      </FormGroup>
-
-                      <FormGroup
-                        advancedSettings={advancedSettings}
-                        isAdvanced={true}
-                      >
-                        <FormLabel>{translate('ChownGroup')}</FormLabel>
-
-                        <FormInputGroup
-                          type={inputTypes.TEXT}
-                          name="chownGroup"
-                          helpText={translate('ChownGroupHelpText')}
-                          helpTextWarning={translate('ChownGroupHelpTextWarning')}
-                          values={fileDateOptions}
-                          onChange={onInputChange}
-                          {...settings.chownGroup}
-                        />
-                      </FormGroup>
-                    </FieldSet> : null
-                }
-              </Form> : null
-          }
-
-          <FieldSet legend={translate('RootFolders')}>
-            <RootFolders />
-            <AddRootFolder />
-          </FieldSet>
-        </PageContentBody>
-      </PageContent>
-    );
-  }
-
-}
-
-MediaManagement.propTypes = {
-  advancedSettings: PropTypes.bool.isRequired,
-  isFetching: PropTypes.bool.isRequired,
-  error: PropTypes.object,
-  settings: PropTypes.object.isRequired,
-  hasSettings: PropTypes.bool.isRequired,
-  isWindows: PropTypes.bool.isRequired,
-  onSavePress: PropTypes.func.isRequired,
-  onInputChange: PropTypes.func.isRequired
-};
-
-export default MediaManagement;
