+++ frontend/src/System/useIsWindows.ts
@@ -1,0 +1,8 @@
+import { useSelector } from 'react-redux';
+import AppState from 'App/State/AppState';
+
+function useIsWindows() {
+  return useSelector((state: AppState) => state.system.status.item.isWindows);
+}
+
+export default useIsWindows;
